Hotfix for scraper crash and scraper refactoring with some tests
