+++ .gitignore
@@ -54,3 +54,5 @@ google-services.json
 freeline.py
 freeline/
 freeline_project_description.json
+
+*.local
