+++ app/build.gradle
@@ -29,12 +29,12 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'org.jetbrains.anko:anko-common:0.9'
-    implementation 'com.android.support:appcompat-v7:27.1.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    implementation 'com.android.support:cardview-v7:27.1.0'
-    implementation 'com.android.support:recyclerview-v7:27.1.0'
+    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.0'
+    implementation 'com.android.support:cardview-v7:27.1.1'
+    implementation 'com.android.support:recyclerview-v7:27.1.1'
     // https://mvnrepository.com/artifact/com.google.code.gson/gson
     implementation 'com.google.code.gson:gson:2.8.2'
     // Koin for Android
@@ -43,7 +43,7 @@ dependencies {
     // jsoup HTML parser library @ https://jsoup.org/
     implementation 'org.jsoup:jsoup:1.11.2'
     implementation 'net.sf.biweekly:biweekly:0.6.1'
-    implementation 'com.android.support:design:27.1.0'
+    implementation 'com.android.support:design:27.1.1'
     //rxJava for submitting the models to views
     implementation 'io.reactivex.rxjava2:rxkotlin:2.2.0'
     testImplementation 'junit:junit:4.12'
@@ -52,8 +52,8 @@ dependencies {
     // Koin testing tools
     testImplementation 'org.koin:koin-test:0.8.1'
     testImplementation 'org.mockito:mockito-core:2.16.0'
-    androidTestCompile 'org.mockito:mockito-android:2.8.9'
-    androidTestCompile('com.nhaarman:mockito-kotlin:1.4.0', {
+    androidTestImplementation 'org.mockito:mockito-android:2.8.9'
+    androidTestImplementation('com.nhaarman:mockito-kotlin:1.4.0', {
         exclude group: 'org.jetbrains.kotlin'
         exclude group: 'org.mockito'
     })
