+++ app/src/main/java/com/ice/ktuice/al/GradeTable/GradeTableManager.kt
@@ -3,20 +3,11 @@ package com.ice.ktuice.al.GradeTable
 import com.ice.ktuice.al.GradeTable.gradeTableModels.GradeTableFactory
 import com.ice.ktuice.al.GradeTable.gradeTableModels.GradeTableModel
 import com.ice.ktuice.al.GradeTable.gradeTableModels.SemesterAdapterItem
-import com.ice.ktuice.DAL.repositories.gradeResponseRepository.YearGradesRepository
-import com.ice.ktuice.DAL.repositories.loginRepository.LoginRepository
-import com.ice.ktuice.DAL.repositories.prefrenceRepository.PreferenceRepository
-import com.ice.ktuice.R
 import com.ice.ktuice.al.services.userService.UserService
 import com.ice.ktuice.al.services.yearGradesService.YearGradesService
-import com.ice.ktuice.models.LoginModel
 import com.ice.ktuice.models.YearGradesCollectionModel
-import com.ice.ktuice.models.YearGradesModel
 import com.ice.ktuice.models.YearModel
-import com.ice.ktuice.scraperService.exceptions.AuthenticationException
-import com.ice.ktuice.scraperService.ScraperService
-import io.realm.RealmList
-import io.realm.RealmResults
+import com.ice.ktuice.scraperService.ktuScraperService.exceptions.AuthenticationException
 import org.jetbrains.anko.getStackTraceString
 import org.koin.standalone.KoinComponent
 import org.koin.standalone.inject
