+++ app/src/main/java/com/ice/ktuice/al/lectureCalendar/CalendarManager.kt
@@ -2,7 +2,7 @@ package com.ice.ktuice.al.lectureCalendar
 
 import com.ice.ktuice.DAL.repositories.calendarRepository.CalendarRepositoryImpl
 import com.ice.ktuice.models.lectureCalendarModels.CalendarModel
-import com.ice.ktuice.scraperService.handlers.CalendarHandler
+import com.ice.ktuice.scraperService.ktuScraperService.handlers.CalendarHandler
 import com.ice.ktuice.al.services.userService.UserService
 import io.reactivex.subjects.ReplaySubject
 import io.reactivex.subjects.Subject
