+++ app/src/main/java/com/ice/ktuice/al/services/userService/UserServiceImpl.kt
@@ -14,6 +14,7 @@ import org.koin.standalone.inject
 class UserServiceImpl: UserService, KoinComponent{
     private val preferenceRepository: PreferenceRepository by inject()
     private val loginRepository: LoginRepository by inject()
+    private val scraperService: ScraperService by inject()
 
     override fun getLoginForCurrentUser(): LoginModel {
         val requestedStudentId = preferenceRepository.getValue(R.string.logged_in_user_code)
@@ -36,7 +37,7 @@ class UserServiceImpl: UserService, KoinComponent{
      */
     override fun refreshLoginCookies(): LoginModel? {
         val loginModel = getLoginForCurrentUser()
-        val newLoginModelResponse = ScraperService.login(loginModel.username, loginModel.password)
+        val newLoginModelResponse = scraperService.login(loginModel.username, loginModel.password)
         val newLoginModel = newLoginModelResponse.loginModel
         loginModel.authCookies.clear()
         loginModel.authCookies.addAll(newLoginModel?.authCookies!!)
