+++ app/src/main/java/com/ice/ktuice/repositories/BaseRepository.kt
@@ -1,0 +1,8 @@
+package com.ice.ktuice.DAL.repositories
+
+import io.realm.Realm
+import io.realm.RealmObject
+
+abstract class BaseRepository<T: RealmObject> {
+    inline fun <reified T: RealmObject>where(realm: Realm) = realm.where(T::class.java)!!
+}
