+++ app/src/main/java/com/ice/ktuice/repositories/calendarRepository/CalendarRepository.kt
@@ -1,0 +1,17 @@
+package com.ice.ktuice.DAL.repositories.calendarRepository
+
+import com.ice.ktuice.models.lectureCalendarModels.CalendarModel
+
+/**
+ * Created by Andrius on 3/10/2018.
+ * Repository to store the lectures for a student
+ */
+interface CalendarRepository {
+
+    fun getByStudCode(studCode: String): CalendarModel?
+
+    /**
+     * Uses the currently logged in student to store the calendar
+     */
+    fun createOrUpdate(calendarModel: CalendarModel)
+}
