--- app/src/main/java/com/ice/ktuice/scraperService/RetroClient.kt
@@ -1,10 +1,0 @@
-package com.ice.ktuice.scraperService
-
-import retrofit2.Retrofit
-
-object RetroClient {
-    const val BASE_URL = "https://uais.cr.ktu.lt"
-    val client = Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .build()!!
-}
