+++ app/src/main/java/com/ice/ktuice/scraperService/ScraperService.kt
@@ -1,54 +1,36 @@
 package com.ice.ktuice.scraperService
 
-import com.ice.ktuice.scraperService.handlers.DataHandler
-import com.ice.ktuice.scraperService.handlers.LoginHandler
 import com.ice.ktuice.models.LoginModel
 import com.ice.ktuice.models.YearGradesModel
 import com.ice.ktuice.models.YearModel
-import com.ice.ktuice.models.responses.YearGradesResponseModel
-import com.ice.ktuice.scraperService.exceptions.AuthenticationException
-import com.ice.ktuice.scraperService.exceptions.ServerErrorException
+import com.ice.ktuice.models.responses.LoginResponse
+import com.ice.ktuice.scraperService.ktuScraperService.exceptions.AuthenticationException
+import com.ice.ktuice.scraperService.ktuScraperService.exceptions.ServerErrorException
 
-
-object ScraperService {
+interface ScraperService {
     /**
      * Gets authentication cookies
      * and general information about a student
      * @return Authentication cookies and user info
      */
-    fun login(username: String, password: String)
-            = LoginHandler().getAuthCookies(username, password)
+    fun login(username: String, password: String): LoginResponse
 
     /**
      * Gets all marks for requested year.
      *
-     * @param authCookie response from loginModel used for authentication.
+     * @param login response from loginModel used for authentication.
      * @param yearModel from loginModel
      * @return List marks
      *
      * @throws AuthenticationException if cookies timeout or are incorrect
      * @throws ServerErrorException if the server responds with code 500
      */
-    fun getGrades(authCookie: LoginModel, yearModel: YearModel): YearGradesModel {
-        var response: YearGradesResponseModel
-        try {
-             response = DataHandler.getGrades(authCookie, yearModel)
-        }catch (e: AuthenticationException){
-            println("Cookies expired! Logging in again!")
-            val newLogin = refreshLoginCookies(authCookie)!!
-            response = DataHandler.getGrades(newLogin, yearModel)
-        }
+    fun getGrades(login: LoginModel, yearModel: YearModel): YearGradesModel
 
-        println("Grades response code(@Scraper service):"+response.statusCode)
-        if(response.statusCode >= 500){
-            throw ServerErrorException("Server error, something went wrong!")
-        }
-        return response.yearGradesModel
-    }
-
-    fun refreshLoginCookies(login: LoginModel): LoginModel?{
-        return login(login.username, login.password).loginModel
-    }
-
+    /**
+     * Used for refreshing cookies (if say fetching the grades fails, the service assumes that
+     * the cookies are expired and tries to refresh the cookies at least once)
+     * works exactly as the login() function, just for convenience
+     */
+    fun refreshLoginCookies(login: LoginModel): LoginResponse
 }
-
