--- app/src/main/java/com/ice/ktuice/scraperService/exceptions/AuthenticationException.kt
@@ -1,7 +1,0 @@
-package com.ice.ktuice.scraperService.exceptions
-
-/**
- * Created by Andrius on 1/31/2018.
- * Exception raised by the auth service when the username or password are not correct
- */
-class AuthenticationException(message: String = ""): Exception(message)
