--- app/src/main/java/com/ice/ktuice/scraperService/exceptions/ServerErrorException.kt
@@ -1,6 +1,0 @@
-package com.ice.ktuice.scraperService.exceptions
-
-/**
- * Created by Andrius on 1/31/2018.
- */
-class ServerErrorException(message: String = ""): Exception(message)
