--- app/src/main/java/com/ice/ktuice/scraperService/handlers/BaseHandler.kt
@@ -1,8 +1,0 @@
-package com.ice.ktuice.scraperService.handlers
-
-import com.ice.ktuice.scraperService.RetroClient
-
-open class BaseHandler {
-    protected fun <T> Class<T>.create(): T
-            = RetroClient.client.create(this)
-}
