+++ app/src/main/java/com/ice/ktuice/scraperService/ktuScraperService/RetroClient.kt
@@ -1,0 +1,10 @@
+package com.ice.ktuice.scraperService.ktuScraperService
+
+import retrofit2.Retrofit
+
+object RetroClient {
+    const val BASE_URL = "https://uais.cr.ktu.lt"
+    val client = Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .build()!!
+}
