+++ app/src/main/java/com/ice/ktuice/scraperService/ktuScraperService/exceptions/ServerErrorException.kt
@@ -1,0 +1,6 @@
+package com.ice.ktuice.scraperService.ktuScraperService.exceptions
+
+/**
+ * Created by Andrius on 1/31/2018.
+ */
+class ServerErrorException(message: String = ""): Exception(message)
