+++ app/src/main/java/com/ice/ktuice/scraperService/ktuScraperService/handlers/BaseHandler.kt
@@ -1,0 +1,8 @@
+package com.ice.ktuice.scraperService.ktuScraperService.handlers
+
+import com.ice.ktuice.scraperService.ktuScraperService.RetroClient
+
+open class BaseHandler {
+    protected fun <T> Class<T>.create(): T
+            = RetroClient.client.create(this)
+}
