+++ app/src/test/java/com/ice/ktuice/YearGradesComparatorTest.kt
@@ -1,7 +1,7 @@
 package com.ice.ktuice
 
-import com.ice.ktuice.TestYearGradesCollectionProvider.Companion.addMark
-import com.ice.ktuice.TestYearGradesCollectionProvider.Companion.createDefaultYearGradesCollection
+import com.ice.ktuice.YearGradesCollectionProviderTest.Companion.addMark
+import com.ice.ktuice.YearGradesCollectionProviderTest.Companion.createDefaultYearGradesCollection
 import com.ice.ktuice.al.GradeTable.yearGradesModelComparator.Difference
 import com.ice.ktuice.al.GradeTable.yearGradesModelComparator.YearGradesModelComparator
 import com.ice.ktuice.al.GradeTable.yearGradesModelComparator.YearGradesModelComparatorImpl
