+++ week-view-lib/build.gradle
@@ -18,6 +18,6 @@ configurations {
     javadocDeps
 }
 dependencies {
-    compile 'com.android.support:appcompat-v7:27.1.0'
+    implementation 'com.android.support:appcompat-v7:27.1.0'
     javadocDeps 'com.android.support:appcompat-v7:27.1.0'
 }
